% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.encryption}
\alias{Bucket.encryption}
\title{Bucket.encryption Object}
\usage{
Bucket.encryption(default_kms_key_name = NULL)
}
\arguments{
\item{default_kms_key_name}{No description}
}
\value{
Bucket.encryption object
}
\description{
Bucket.encryption Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encryption configuration used by default for newly inserted objects, when no encryption config is specified.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.logging}}, \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}

