% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.lifecycle}
\alias{Bucket.lifecycle}
\title{Bucket.lifecycle Object}
\usage{
Bucket.lifecycle(Bucket.lifecycle.rule = NULL,
  Bucket.lifecycle.rule.action = NULL,
  Bucket.lifecycle.rule.condition = NULL, rule = NULL)
}
\arguments{
\item{Bucket.lifecycle.rule}{The \link{Bucket.lifecycle.rule} object or list of objects}

\item{Bucket.lifecycle.rule.action}{The \link{Bucket.lifecycle.rule.action} object or list of objects}

\item{Bucket.lifecycle.rule.condition}{The \link{Bucket.lifecycle.rule.condition} object or list of objects}

\item{rule}{A lifecycle management rule, which is made of an action to take and the condition(s) under which the action will be taken}
}
\value{
Bucket.lifecycle object
}
\description{
Bucket.lifecycle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The bucket's lifecycle configuration. See lifecycle management for more information.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.logging}}, \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}
