% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.lifecycle.rule.action}
\alias{Bucket.lifecycle.rule.action}
\title{Bucket.lifecycle.rule.action Object}
\usage{
Bucket.lifecycle.rule.action(type = NULL)
}
\arguments{
\item{type}{Type of the action}
}
\value{
Bucket.lifecycle.rule.action object
}
\description{
Bucket.lifecycle.rule.action Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The action to take.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.encryption}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.logging}}, \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}

