% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.lifecycle.rule.condition}
\alias{Bucket.lifecycle.rule.condition}
\title{Bucket.lifecycle.rule.condition Object}
\usage{
Bucket.lifecycle.rule.condition(age = NULL, createdBefore = NULL,
  isLive = NULL, matchesStorageClass = NULL, numNewerVersions = NULL)
}
\arguments{
\item{age}{Age of an object (in days)}

\item{createdBefore}{A date in RFC 3339 format with only the date part (for instance, '2013-01-15')}

\item{isLive}{Relevant only for versioned objects}

\item{matchesStorageClass}{Objects having any of the storage classes specified by this condition will be matched}

\item{numNewerVersions}{Relevant only for versioned objects}
}
\value{
Bucket.lifecycle.rule.condition object
}
\description{
Bucket.lifecycle.rule.condition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The condition(s) under which the action will be taken.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.logging}}, \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}
