% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.lifecycle.rule}
\alias{Bucket.lifecycle.rule}
\title{Bucket.lifecycle.rule Object}
\usage{
Bucket.lifecycle.rule(Bucket.lifecycle.rule.action = NULL,
  Bucket.lifecycle.rule.condition = NULL)
}
\arguments{
\item{Bucket.lifecycle.rule.action}{The \link{Bucket.lifecycle.rule.action} object or list of objects}

\item{Bucket.lifecycle.rule.condition}{The \link{Bucket.lifecycle.rule.condition} object or list of objects}
}
\value{
Bucket.lifecycle.rule object
}
\description{
Bucket.lifecycle.rule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.encryption}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.logging}}, \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}

