% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.logging}
\alias{Bucket.logging}
\title{Bucket.logging Object}
\usage{
Bucket.logging(logBucket = NULL, logObjectPrefix = NULL)
}
\arguments{
\item{logBucket}{The destination bucket where the current bucket's logs should be placed}

\item{logObjectPrefix}{A prefix for log object names}
}
\value{
Bucket.logging object
}
\description{
Bucket.logging Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.encryption}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}

