% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket}
\alias{Bucket}
\title{Cloud Storage JSON API Objects 
Stores and retrieves potentially large, immutable data objects.}
\usage{
Bucket(Bucket.cors = NULL, Bucket.lifecycle = NULL,
  Bucket.lifecycle.rule = NULL, Bucket.lifecycle.rule.action = NULL,
  Bucket.lifecycle.rule.condition = NULL, Bucket.logging = NULL,
  Bucket.owner = NULL, Bucket.versioning = NULL, Bucket.website = NULL,
  acl = NULL, cors = NULL, defaultObjectAcl = NULL, etag = NULL,
  id = NULL, lifecycle = NULL, location = NULL, logging = NULL,
  metageneration = NULL, name = NULL, owner = NULL,
  projectNumber = NULL, selfLink = NULL, storageClass = NULL,
  timeCreated = NULL, updated = NULL, versioning = NULL, website = NULL)
}
\arguments{
\item{Bucket.cors}{The \link{Bucket.cors} object or list of objects}

\item{Bucket.lifecycle}{The \link{Bucket.lifecycle} object or list of objects}

\item{Bucket.lifecycle.rule}{The \link{Bucket.lifecycle.rule} object or list of objects}

\item{Bucket.lifecycle.rule.action}{The \link{Bucket.lifecycle.rule.action} object or list of objects}

\item{Bucket.lifecycle.rule.condition}{The \link{Bucket.lifecycle.rule.condition} object or list of objects}

\item{Bucket.logging}{The \link{Bucket.logging} object or list of objects}

\item{Bucket.owner}{The \link{Bucket.owner} object or list of objects}

\item{Bucket.versioning}{The \link{Bucket.versioning} object or list of objects}

\item{Bucket.website}{The \link{Bucket.website} object or list of objects}

\item{acl}{Access controls on the bucket}

\item{cors}{The bucket's Cross-Origin Resource Sharing (CORS) configuration}

\item{defaultObjectAcl}{Default access controls to apply to new objects when no ACL is provided}

\item{etag}{HTTP 1}

\item{id}{The ID of the bucket}

\item{lifecycle}{The bucket's lifecycle configuration}

\item{location}{The location of the bucket}

\item{logging}{The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs}

\item{metageneration}{The metadata generation of this bucket}

\item{name}{The name of the bucket}

\item{owner}{The owner of the bucket}

\item{projectNumber}{The project number of the project the bucket belongs to}

\item{selfLink}{The URI of this bucket}

\item{storageClass}{The bucket's default storage class, used whenever no storageClass is specified for a newly-created object}

\item{timeCreated}{The creation time of the bucket in RFC 3339 format}

\item{updated}{The modification time of the bucket in RFC 3339 format}

\item{versioning}{The bucket's versioning configuration}

\item{website}{The bucket's website configuration, controlling how the service behaves when accessing bucket contents as a web site}
}
\value{
Bucket object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:17:01
filename: /Users/mark/dev/R/autoGoogleAPI/googlestoragev1.auto/R/storage_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Bucket Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A bucket.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.logging}}, \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}
