% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.website}
\alias{Bucket.website}
\title{Bucket.website Object}
\usage{
Bucket.website(mainPageSuffix = NULL, notFoundPage = NULL)
}
\arguments{
\item{mainPageSuffix}{If the requested object path is missing, the service will ensure the path has a trailing '/', append this suffix, and attempt to retrieve the resulting object}

\item{notFoundPage}{If the requested object path is missing, and any mainPageSuffix object is missing, if applicable, the service will return the named object from this bucket as the content for a 404 Not Found result}
}
\value{
Bucket.website object
}
\description{
Bucket.website Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The bucket's website configuration, controlling how the service behaves when accessing bucket contents as a web site. See the Static Website Examples for more information.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.logging}}, \code{\link{Bucket.owner}},
  \code{\link{Bucket.versioning}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}
