% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{BucketAccessControl.projectTeam}
\alias{BucketAccessControl.projectTeam}
\title{BucketAccessControl.projectTeam Object}
\usage{
BucketAccessControl.projectTeam(projectNumber = NULL, team = NULL)
}
\arguments{
\item{projectNumber}{The project number}

\item{team}{The team}
}
\value{
BucketAccessControl.projectTeam object
}
\description{
BucketAccessControl.projectTeam Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The project team associated with the entity, if any.
}
\seealso{
Other BucketAccessControl functions: \code{\link{BucketAccessControl}},
  \code{\link{bucketAccessControls.insert}},
  \code{\link{bucketAccessControls.patch}},
  \code{\link{bucketAccessControls.update}}
}

