% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{ComposeRequest}
\alias{ComposeRequest}
\title{ComposeRequest Object}
\usage{
ComposeRequest(ComposeRequest.sourceObjects = NULL,
  ComposeRequest.sourceObjects.objectPreconditions = NULL,
  destination = NULL, sourceObjects = NULL)
}
\arguments{
\item{ComposeRequest.sourceObjects}{The \link{ComposeRequest.sourceObjects} object or list of objects}

\item{ComposeRequest.sourceObjects.objectPreconditions}{The \link{ComposeRequest.sourceObjects.objectPreconditions} object or list of objects}

\item{destination}{Properties of the resulting object}

\item{sourceObjects}{The list of source objects that will be concatenated into a single object}
}
\value{
ComposeRequest object
}
\description{
ComposeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Compose request.
}
\seealso{
Other ComposeRequest functions: \code{\link{ComposeRequest.sourceObjects.objectPreconditions}},
  \code{\link{ComposeRequest.sourceObjects}},
  \code{\link{objects.compose}}
}

