% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{ComposeRequest.sourceObjects.objectPreconditions}
\alias{ComposeRequest.sourceObjects.objectPreconditions}
\title{ComposeRequest.sourceObjects.objectPreconditions Object}
\usage{
ComposeRequest.sourceObjects.objectPreconditions(ifGenerationMatch = NULL)
}
\arguments{
\item{ifGenerationMatch}{Only perform the composition if the generation of the source object that would be used matches this value}
}
\value{
ComposeRequest.sourceObjects.objectPreconditions object
}
\description{
ComposeRequest.sourceObjects.objectPreconditions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Conditions that must be met for this operation to execute.
}
\seealso{
Other ComposeRequest functions: \code{\link{ComposeRequest.sourceObjects}},
  \code{\link{ComposeRequest}},
  \code{\link{objects.compose}}
}

