% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{ComposeRequest.sourceObjects}
\alias{ComposeRequest.sourceObjects}
\title{ComposeRequest.sourceObjects Object}
\usage{

  ComposeRequest.sourceObjects(ComposeRequest.sourceObjects.objectPreconditions = NULL)
}
\arguments{
\item{ComposeRequest.sourceObjects.objectPreconditions}{The \link{ComposeRequest.sourceObjects.objectPreconditions} object or list of objects}
}
\value{
ComposeRequest.sourceObjects object
}
\description{
ComposeRequest.sourceObjects Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ComposeRequest functions: \code{\link{ComposeRequest.sourceObjects.objectPreconditions}},
  \code{\link{ComposeRequest}},
  \code{\link{objects.compose}}
}

