% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{defaultObjectAccessControls.list}
\alias{defaultObjectAccessControls.list}
\title{Retrieves default object ACL entries on the specified bucket.}
\usage{
defaultObjectAccessControls.list(bucket, ifMetagenerationMatch = NULL,
  ifMetagenerationNotMatch = NULL)
}
\arguments{
\item{bucket}{Name of a bucket}

\item{ifMetagenerationMatch}{If present, only return default ACL listing if the bucket's current metageneration matches this value}

\item{ifMetagenerationNotMatch}{If present, only return default ACL listing if the bucket's current metageneration does not match the given value}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.full_control
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.full_control)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}
}

