% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{defaultObjectAccessControls.update}
\alias{defaultObjectAccessControls.update}
\title{Updates a default object ACL entry on the specified bucket.}
\usage{
defaultObjectAccessControls.update(ObjectAccessControl, bucket, entity)
}
\arguments{
\item{ObjectAccessControl}{The \link{ObjectAccessControl} object to pass to this method}

\item{bucket}{Name of a bucket}

\item{entity}{The entity holding the permission}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.full_control
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.full_control)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other ObjectAccessControl functions: \code{\link{ObjectAccessControl.projectTeam}},
  \code{\link{ObjectAccessControl}},
  \code{\link{defaultObjectAccessControls.insert}},
  \code{\link{defaultObjectAccessControls.patch}},
  \code{\link{objectAccessControls.insert}},
  \code{\link{objectAccessControls.patch}},
  \code{\link{objectAccessControls.update}}
}

