% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Object.customerEncryption}
\alias{Object.customerEncryption}
\title{Object.customerEncryption Object}
\usage{
Object.customerEncryption(encryptionAlgorithm = NULL, keySha256 = NULL)
}
\arguments{
\item{encryptionAlgorithm}{The encryption algorithm}

\item{keySha256}{SHA256 hash value of the encryption key}
}
\value{
Object.customerEncryption object
}
\description{
Object.customerEncryption Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata of customer-supplied encryption key, if the object is encrypted by such a key.
}
\seealso{
Other Object functions: \code{\link{Object.metadata}},
  \code{\link{Object.owner}}, \code{\link{Object}},
  \code{\link{objects.copy}}, \code{\link{objects.insert}},
  \code{\link{objects.patch}},
  \code{\link{objects.rewrite}},
  \code{\link{objects.update}}
}

