% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Object.metadata}
\alias{Object.metadata}
\title{Object.metadata Object}
\usage{
Object.metadata()
}
\value{
Object.metadata object
}
\description{
Object.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User-provided metadata, in key/value pairs.
}
\seealso{
Other Object functions: \code{\link{Object.customerEncryption}},
  \code{\link{Object.owner}}, \code{\link{Object}},
  \code{\link{objects.copy}}, \code{\link{objects.insert}},
  \code{\link{objects.patch}},
  \code{\link{objects.rewrite}},
  \code{\link{objects.update}}
}

