% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Object.owner}
\alias{Object.owner}
\title{Object.owner Object}
\usage{
Object.owner(entity = NULL, entityId = NULL)
}
\arguments{
\item{entity}{The entity, in the form user-userId}

\item{entityId}{The ID for the entity}
}
\value{
Object.owner object
}
\description{
Object.owner Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The owner of the object. This will always be the uploader of the object.
}
\seealso{
Other Object functions: \code{\link{Object.customerEncryption}},
  \code{\link{Object.metadata}}, \code{\link{Object}},
  \code{\link{objects.copy}}, \code{\link{objects.insert}},
  \code{\link{objects.patch}},
  \code{\link{objects.rewrite}},
  \code{\link{objects.update}}
}
