% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{ObjectAccessControl.projectTeam}
\alias{ObjectAccessControl.projectTeam}
\title{ObjectAccessControl.projectTeam Object}
\usage{
ObjectAccessControl.projectTeam(projectNumber = NULL, team = NULL)
}
\arguments{
\item{projectNumber}{The project number}

\item{team}{The team}
}
\value{
ObjectAccessControl.projectTeam object
}
\description{
ObjectAccessControl.projectTeam Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The project team associated with the entity, if any.
}
\seealso{
Other ObjectAccessControl functions: \code{\link{ObjectAccessControl}},
  \code{\link{defaultObjectAccessControls.insert}},
  \code{\link{defaultObjectAccessControls.patch}},
  \code{\link{defaultObjectAccessControls.update}},
  \code{\link{objectAccessControls.insert}},
  \code{\link{objectAccessControls.patch}},
  \code{\link{objectAccessControls.update}}
}

