% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{ObjectAccessControl}
\alias{ObjectAccessControl}
\title{ObjectAccessControl Object}
\usage{
ObjectAccessControl(ObjectAccessControl.projectTeam = NULL, bucket = NULL,
  domain = NULL, email = NULL, entity = NULL, entityId = NULL,
  etag = NULL, generation = NULL, id = NULL, object = NULL,
  projectTeam = NULL, role = NULL, selfLink = NULL)
}
\arguments{
\item{ObjectAccessControl.projectTeam}{The \link{ObjectAccessControl.projectTeam} object or list of objects}

\item{bucket}{The name of the bucket}

\item{domain}{The domain associated with the entity, if any}

\item{email}{The email address associated with the entity, if any}

\item{entity}{The entity holding the permission, in one of the following forms:}

\item{entityId}{The ID for the entity, if any}

\item{etag}{HTTP 1}

\item{generation}{The content generation of the object}

\item{id}{The ID of the access-control entry}

\item{object}{The name of the object}

\item{projectTeam}{The project team associated with the entity, if any}

\item{role}{The access permission for the entity}

\item{selfLink}{The link to this access-control entry}
}
\value{
ObjectAccessControl object
}
\description{
ObjectAccessControl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An access-control entry.
}
\seealso{
Other ObjectAccessControl functions: \code{\link{ObjectAccessControl.projectTeam}},
  \code{\link{defaultObjectAccessControls.insert}},
  \code{\link{defaultObjectAccessControls.patch}},
  \code{\link{defaultObjectAccessControls.update}},
  \code{\link{objectAccessControls.insert}},
  \code{\link{objectAccessControls.patch}},
  \code{\link{objectAccessControls.update}}
}

