% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{ObjectAccessControls}
\alias{ObjectAccessControls}
\title{ObjectAccessControls Object}
\usage{
ObjectAccessControls(items = NULL)
}
\arguments{
\item{items}{The list of items}
}
\value{
ObjectAccessControls object
}
\description{
ObjectAccessControls Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An access-control list.
}

