% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{objects.compose}
\alias{objects.compose}
\title{Concatenates a list of existing objects into a new object in the same bucket.}
\usage{
objects.compose(ComposeRequest, destinationBucket, destinationObject,
  destinationPredefinedAcl = NULL, ifGenerationMatch = NULL,
  ifMetagenerationMatch = NULL, kmsKeyName = NULL)
}
\arguments{
\item{ComposeRequest}{The \link{ComposeRequest} object to pass to this method}

\item{destinationBucket}{Name of the bucket in which to store the new object}

\item{destinationObject}{Name of the new object}

\item{destinationPredefinedAcl}{Apply a predefined set of access controls to the destination object}

\item{ifGenerationMatch}{Makes the operation conditional on whether the object's current generation matches the given value}

\item{ifMetagenerationMatch}{Makes the operation conditional on whether the object's current metageneration matches the given value}

\item{kmsKeyName}{Resource name of the Cloud KMS key, of the form projects/my-project/locations/global/keyRings/my-kr/cryptoKeys/my-key, that will be used to encrypt the object}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other ComposeRequest functions: \code{\link{ComposeRequest.sourceObjects.objectPreconditions}},
  \code{\link{ComposeRequest.sourceObjects}},
  \code{\link{ComposeRequest}}
}

