% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{objects.rewrite}
\alias{objects.rewrite}
\title{Rewrites a source object to a destination object. Optionally overrides metadata.}
\usage{
objects.rewrite(Object, sourceBucket, sourceObject, destinationBucket,
  destinationObject, destinationKmsKeyName = NULL,
  destinationPredefinedAcl = NULL, ifGenerationMatch = NULL,
  ifGenerationNotMatch = NULL, ifMetagenerationMatch = NULL,
  ifMetagenerationNotMatch = NULL, ifSourceGenerationMatch = NULL,
  ifSourceGenerationNotMatch = NULL, ifSourceMetagenerationMatch = NULL,
  ifSourceMetagenerationNotMatch = NULL, maxBytesRewrittenPerCall = NULL,
  projection = NULL, rewriteToken = NULL, sourceGeneration = NULL)
}
\arguments{
\item{Object}{The \link{Object} object to pass to this method}

\item{sourceBucket}{Name of the bucket in which to find the source object}

\item{sourceObject}{Name of the source object}

\item{destinationBucket}{Name of the bucket in which to store the new object}

\item{destinationObject}{Name of the new object}

\item{destinationKmsKeyName}{Resource name of the Cloud KMS key, of the form projects/my-project/locations/global/keyRings/my-kr/cryptoKeys/my-key, that will be used to encrypt the object}

\item{destinationPredefinedAcl}{Apply a predefined set of access controls to the destination object}

\item{ifGenerationMatch}{Makes the operation conditional on whether the destination object's current generation matches the given value}

\item{ifGenerationNotMatch}{Makes the operation conditional on whether the destination object's current generation does not match the given value}

\item{ifMetagenerationMatch}{Makes the operation conditional on whether the destination object's current metageneration matches the given value}

\item{ifMetagenerationNotMatch}{Makes the operation conditional on whether the destination object's current metageneration does not match the given value}

\item{ifSourceGenerationMatch}{Makes the operation conditional on whether the source object's generation matches the given value}

\item{ifSourceGenerationNotMatch}{Makes the operation conditional on whether the source object's generation does not match the given value}

\item{ifSourceMetagenerationMatch}{Makes the operation conditional on whether the source object's current metageneration matches the given value}

\item{ifSourceMetagenerationNotMatch}{Makes the operation conditional on whether the source object's current metageneration does not match the given value}

\item{maxBytesRewrittenPerCall}{The maximum number of bytes that will be rewritten per rewrite request}

\item{projection}{Set of properties to return}

\item{rewriteToken}{Include this field (from the previous rewrite response) on each rewrite request after the first one, until the rewrite response 'done' flag is true}

\item{sourceGeneration}{If present, selects a specific revision of the source object (as opposed to the latest version, the default)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other Object functions: \code{\link{Object.customerEncryption}},
  \code{\link{Object.metadata}},
  \code{\link{Object.owner}}, \code{\link{Object}},
  \code{\link{objects.copy}}, \code{\link{objects.insert}},
  \code{\link{objects.patch}}, \code{\link{objects.update}}
}

