% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{objects.watchAll}
\alias{objects.watchAll}
\title{Watch for changes on all objects in a bucket.}
\usage{
objects.watchAll(Channel, bucket, delimiter = NULL, maxResults = NULL,
  pageToken = NULL, prefix = NULL, projection = NULL, versions = NULL)
}
\arguments{
\item{Channel}{The \link{Channel} object to pass to this method}

\item{bucket}{Name of the bucket in which to look for objects}

\item{delimiter}{Returns results in a directory-like mode}

\item{maxResults}{Maximum number of items plus prefixes to return}

\item{pageToken}{A previously-returned page token representing part of the larger set of results to view}

\item{prefix}{Filter results to objects whose names begin with this prefix}

\item{projection}{Set of properties to return}

\item{versions}{If true, lists all versions of an object as distinct results}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_only, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other Channel functions: \code{\link{Channel.params}},
  \code{\link{Channel}}, \code{\link{channels.stop}}
}

