% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{RewriteResponse}
\alias{RewriteResponse}
\title{RewriteResponse Object}
\usage{
RewriteResponse(done = NULL, objectSize = NULL, resource = NULL,
  rewriteToken = NULL, totalBytesRewritten = NULL)
}
\arguments{
\item{done}{true if the copy is finished; otherwise, false if the copy is in progress}

\item{objectSize}{The total size of the object being copied in bytes}

\item{resource}{A resource containing the metadata for the copied-to object}

\item{rewriteToken}{A token to use in subsequent requests to continue copying data}

\item{totalBytesRewritten}{The total bytes written so far, which can be used to provide a waiting user with a progress indicator}
}
\value{
RewriteResponse object
}
\description{
RewriteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rewrite response.
}
