% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\docType{package}
\name{storage_googleAuthR}
\alias{storage_googleAuthR}
\alias{storage_googleAuthR-package}
\title{Cloud Storage JSON API
Stores and retrieves potentially large, immutable data objects.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:17:00
filename: /Users/mark/dev/R/autoGoogleAPI/googlestoragev1.auto/R/storage_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}
}
